+++ test/goalc/test_with_game.cpp
@@ -337,7 +337,7 @@ TEST_F(WithGameTests, FancyStatic) {
 TEST_F(WithGameTests, IntegerBoxedArray) {
   runner.run_static_test(
       env, testCategory, "test-integer-boxed-array.gc",
-      {"0 0  1 2  2 4  3 6  4 8  5 10  6 12  7 14  8 16  9 18  10 20  11 22  12 40 6\n0\n"});
+      {"0 0  1 2  2 4  3 6  4 8  5 10  6 12  7 14  8 16  9 18  10 20  11 22  12 40 6 array\n0\n"});
 }
 
 TEST_F(WithGameTests, StaticBoxedArray) {